MIDI library extracted.
